Create `openmaps_auth.osm` namespace and remove `openmaps_auth.signals`.
