+++ openmaps_auth/osm/oauth.py
@@ -1,44 +1,17 @@
-import collections
 import logging
 import os.path
 import urllib.parse
-import xml.dom.minidom
 
 import bs4
 import requests
 from django.conf import settings
 from requests_oauthlib import OAuth1Session
 
+from . import api
+from .login import login
+
 
 logger = logging.getLogger(__name__)
-
-
-OSMLogin = collections.namedtuple(
-    "OSMLogin", ["authenticity_token", "cookies", "login_response", "osm_session"]
-)
-
-
-def api_user_details(user):
-    """
-    Returns a dictionary of the results of a user details OpenStreetMap API call.
-    """
-    details_resp = requests.get(
-        settings.OSM_USER_DETAILS_URL, auth=(user.email, user.osm_password)
-    )
-    try:
-        dom = xml.dom.minidom.parseString(details_resp.content)
-    except ValueError:
-        logger.warn("could not parse user details xml")
-        return None
-    user = dom.getElementsByTagName("user")[0]
-    return {
-        "account_created": user.getAttribute("account_created"),
-        "country": user.getAttribute("country"),
-        "display_name": user.getAttribute("display_name"),
-        "email": user.getAttribute("email"),
-        "full_name": user.getAttribute("full_name"),
-        "id": user.getAttribute("id"),
-    }
 
 
 def get_authenticity_token(content):
@@ -47,45 +20,6 @@ def get_authenticity_token(content):
         .find("input", {"name": "authenticity_token"})
         .get("value")
     )
-
-
-def get_csrf_token(content):
-    return (
-        bs4.BeautifulSoup(content, features="html.parser")
-        .find("meta", {"name": "csrf-token"})
-        .get("content")
-    )
-
-
-def login(user):
-    """
-    Performs a login against an OpenStreetMap instance, returning a 4-tuple of
-    an authenticity token string, an initial request cookies object, the
-    login response object, and an OSM session identifier string.
-    """
-    # Generate OSM session.
-    response = requests.get(settings.OSM_BASE_URL)
-    cookies = response.cookies
-    osm_session = cookies[settings.OSM_SESSION_KEY]
-    logger.debug(f"osm session for {user}: {osm_session}")
-
-    # Get OSM CSRF token.
-    response = requests.get(settings.OSM_LOGIN_URL, cookies=cookies)
-    authenticity_token = get_csrf_token(response.content)
-
-    # Login to OSM.
-    login_data = {
-        "referer": "",
-        "authenticity_token": authenticity_token,
-        "username": user.email,
-        "commit": "Login",
-        "password": user.osm_password,
-    }
-    login_response = requests.post(
-        settings.OSM_LOGIN_URL, allow_redirects=False, cookies=cookies, data=login_data
-    )
-
-    return OSMLogin(authenticity_token, cookies, login_response, osm_session)
 
 
 def oauth1_application(
@@ -110,7 +44,7 @@ def oauth1_application(
     ol = login(user)
 
     # oauth1 client applications
-    user_details = api_user_details(user)
+    user_details = api.user_details(user)
     display_name = user_details["display_name"]
     clients_url = f"{settings.OSM_BASE_URL}/user/{display_name}/oauth_clients"
     logger.debug(f"clients_url: {clients_url}")
